telegram/files-downloader/pyproject.toml
[project]
name = "files-downloader"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Amir Aref", email = "amiraref808@gmail.com" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "pyrogram>=2.0.106",
    "tgcrypto>=1.2.5",
]
readme = "README.md"
requires-python = ">= 3.11"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
